requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
toml==0.10.2
twilio==6.54.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
